Add ability to save resume as a PDF
Proof of concept ability to save resume as a pdf. Uses the html2canvas
library to convert the div containing the resume to a canvas element,
and then uses the jsPDF library to convert the canvas to a pdf and save
it. Still need to work on saving at a better resolution.