Make app responsive
The resume now has two fixed sizes. One for devices over 600px, and one
for less. I use mediaqueries extensively to:
1 - Dynamically change the fontsize in the theme
2 - Manually adjust padding, margins, icons sizes.
3 - In some cases, like the chips, manually go in and change properties

Finally, when outputting PDF, change the scaling based on size of resume.

This all feels a little inelegant, however it's important for the resume to
maintain it's relative positioning and sizing so it looks good when
converted to pdf.