REFACTOR PART 1
The begining of a major refactor of the project.
I've lifted all resume updates out to a specfic class that specializes in
updating the resume. Eventually this will live in a useRef in a custom hook.
Updating, adding, and deleting from the resume is handled in a more
functional way.

Theme has been moved out of App

Project is stable after this update.