Create useFirebase hook
Currently the useFirebase hook takes 1 argument, a authStateChangeObserver
function. It returns a signIn, signOut function, as well as a getProfilePic
and getUserName function.

Now, in the appHeader, their is an Avater component displayed. Clicking
on it will allow the user to signIn/signOut. Once signed in the profile
pic is displayed.